pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "hexlet project task manager"
authors = ["pavelkoA <pavelko88@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
dj-database-url = "^2.2.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
uvicorn = "^0.30.1"
flake8 = "^7.0.0"
django-bootstrap5 = "^24.2"
django-filter = "^24.2"
rollbar = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
